WIP: modular capture button improvements / address comments * move capture button component into its own file * decouple the "capture button center" from the capture button and switch component
